fast-api-backend/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.38.0
google-generativeai==0.8.3
pydantic==2.9.2

